Maintenance #CHANGE: Moved BatchInfo class from Resources to Drawing namespace. #CHANGE: Split up a lot of classes into separate files in the EditorBase plugin. #REMOVE: Removed ResourceXy.None and ResourceXy.FileExt constants.
