+++ Duality/Components/Camera.cs
@@ -287,7 +287,7 @@ namespace Duality.Components
 			{
 				this.RenderSinglePass(viewportRect, t);
 			}
-			RenderTarget.Bind(RenderTarget.None);
+			RenderTarget.Bind(null);
 			this.drawDevice.VisibilityMask = this.visibilityMask;
 			this.drawDevice.RenderMode = RenderMatrix.PerspectiveWorld;
 			this.drawDevice.UpdateMatrices(); // Reset matrices for projection calculations during update
@@ -327,7 +327,7 @@ namespace Duality.Components
 				this.drawDevice.Render(ClearFlag.All, ColorRgba.Black, 1.0f);
 				this.drawDevice.PickingIndex = 0;
 
-				RenderTarget.Bind(RenderTarget.None);
+				RenderTarget.Bind(null);
 			}
 
 			// Move data to local buffer
