+++ DualityEditorPlugins/DynamicLighting/DynamicLighting.csproj
@@ -42,6 +42,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DataConverters\LightingRendererFromMaterial.cs" />
     <Compile Include="DynamicLightingPlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
